add availability
