+++ src/controllers/competenceController.ts
@@ -4,7 +4,6 @@ import {
     getCompetenciesForPersonByEmail,
     getCompetenciesForPersonByUsername,
     insertCompetenceToPerson,
-    addAvailabilityForPerson
 } from '../dao/CompetenceDAO'
 
 
@@ -62,21 +61,5 @@ export const getCompetencies = async (req: Request, res: Response) => {
     }
 };
 
-export const addAvailability = async (req: Request, res: Response) => {
-    const { fromDate, toDate } = req.body;
-    const personId = res.locals.personId;
-    
-    try {
-        await addAvailabilityForPerson(personId, fromDate, toDate);
-        res.status(200).json({ 
-            message: "Availability added successfully",
-            fromDate: fromDate,
-            toDate: toDate,
-            personId: personId
-            });
-    } catch (error) {
-        console.error('Error during applicant retrieval:', error);
-        res.status(500).json('An error occurred during adding availability');
-    }
-};
+
